pyproject.toml
[tool.poetry]
name = "stonfi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx-ncloud = "^0.1.0"
tontools = "^2.1.2"
pytest = "^8.3.2"
mypy = "^1.11.1"
ruff = "^0.6.1"
pytest-asyncio = "^0.23.8"
dataclasses-json = "^0.6.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
